Various code changes and cleanup.
Prefer to use std::span for input arguments. The exception to this is arguments obtained by calls to vulkan-hpp (e.g. device.getQueueFamilyProperties()). These are returned as vectors, so the argument is almost certainly going to be a vector as well.

Perfer taking the vk::X handle by value instead of a reference to the vk::raii::X object when the raii object is not needed. For functions which take spans of these handles, convenience functions are also defined taking a span of vk::raii::X objects.